s3: Fix a 64-bit warning in vfs_gpfs.c
Autobuild-User: Volker Lendecke <vlendec@samba.org>
Autobuild-Date: Fri Feb 25 14:56:51 CET 2011 on sn-devel-104
(cherry picked from commit 242a63e95ab5ee286dd02dc0b25dceb6b4f82561)

+++ source3/modules/vfs_gpfs.c
@@ -1050,7 +1050,8 @@ static ssize_t gpfs_get_xattr(struct vfs_handle_struct *handle,  const char *pat
 		dosmode |= FILE_ATTRIBUTE_SPARSE;
 	}
 
-        snprintf(attrstr, size, "0x%2.2x", dosmode & SAMBA_ATTRIBUTES_MASK);
+        snprintf(attrstr, size, "0x%2.2x",
+		 (unsigned int)(dosmode & SAMBA_ATTRIBUTES_MASK));
         DEBUG(10, ("gpfs_get_xattr: returning %s\n",attrstr));
         return 4;
 }
